I changed A to S on many places as they are used for the same thing but A was not defined.
I moved the closed bay stuff up to the bay shapes. You should talk about having a wall vs not having a wall when we talk about the bay shapes.

For m=2 (not m=1) case use the words linear superposition. Also mention that the lambda_n's are positive.

I also cut out the slides on the Trapozodial bay-- we will be tight on time (less then a minute per slide).